test(messages): fold the attachment-parsing tests into the module's own file
A second test file for one source module cut against the repo's convention and
would have read as an oversight. attachment-upload-core.test.ts was only 148
lines, so there was no size argument for splitting it.
